Optimize blend_transformed_tiled_argb().
Profiling indicates that this function is one of two hot spots causing
a noticeable latency when changing KDE virtual desktops.

Instead of computing two modulos per pixel in the inner loop, it is
possible to compute the modulos outside the loop and compute a modulo
sum in the inner loop for a reasonable speedup.

Change-Id: Ic4217b7686e031d7673b3e10aa977dae263096dc
Reviewed-by: Gunnar Sletta <gunnar.sletta@jollamobile.com>