Doc: Added \qtvariable command to QDoc.
Specifies the QT variable needed in the .pro file.
The argument of the command is the qmake QT variable.

To use, add "\qtvariable <value>" to a QDoc comment
which contains the \module command.
QDoc will then associate the class with the QT variable.

Only supported for C++ classes at the moment.
Part of work done for QTBUG-32172

Task-number: QTBUG-32172
Change-Id: Ia8eea30fcfc771191c23a5f5994a48732959ea49
Reviewed-by: Martin Smith <martin.smith@digia.com>