Android: Fix crash when compiling release mode for armv5
When building in thumb mode for armv5 applications will crash
with SIGILL on startup. This has been observed on armv7 devices
and emulators. It could be a bug in the gcc 4.4.3 toolchain, but
since the other toolchains in the NDK have other bugs that
make it impossible to use them for building, we need to disable
thumb until the cross-compiler has been fixed.

Task-number: QTBUG-31338
Change-Id: I22dd228158ef8c43b0b1d6e549d5725c1930536b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>