Enable qsrand() on builds without thread-safe posix
The #ifdef clause in qsrand() needs to be the same as in
qrand(). Otherwise, we will store the seed in thread-local
storage in qsrand(), never passing it into srand(), and then
we'll use regular rand() because the rand_r() function
is missing, thus always using a random seed of 1 in all
applications.

Task-number: QTBUG-32781
Change-Id: I00240a1954ae746b87b031f3a0470a6cbe747571
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>